create subpackage for annotation tests
